requirements.txt
python-telegram-bot==21.5
shodan==1.31.0
python-dotenv==1.0.1
aiohttp==3.10.5
